Add files via upload
changed retbytes() to use strtoul() from stroll(). strtoul() returns an unsigned long - uint64_t is defined as an unsigned long when compiled 64bit.
changed retbytes() to take a pointer to uint64_t rather than return the value. Now returns 1 on success or 0 on failure to get the field or update the pointer.